requirements.txt
Django==1.6.1
Pillow==2.3.0
amqp==1.4.2
anyjson==0.3.3
billiard==3.3.0.14
celery==3.1.8
django-debug-toolbar==1.0.1
django-websocket-redis==0.2.3
gevent==1.0
greenlet==0.4.2
kombu==3.0.10
pytz==2013.9
redis==2.9.1
sqlparse==0.1.10
uWSGI==2.0
wsgiref==0.1.2
